+++ lib/openapi_parser/spec_validator.rb
@@ -17,6 +17,7 @@ require_relative 'spec_validator/rules/dynamic_ref_in_30'
 require_relative 'spec_validator/rules/dynamic_anchor_in_30'
 require_relative 'spec_validator/rules/content_schema_in_30'
 require_relative 'spec_validator/rules/self_before_32'
+require_relative 'spec_validator/rules/tag_fields_before_32'
 
 module OpenAPIParser
   class SpecViolationError < OpenAPIError
@@ -81,6 +82,7 @@ module OpenAPIParser
           Rules::DynamicAnchorIn30,
           Rules::ContentSchemaIn30,
           Rules::SelfBefore32,
+          Rules::TagFieldsBefore32,
         ]
       end
   end